requirements.txt
streamlit>=1.20
sentence-transformers>=2.2.2
transformers>=4.30
torch==2.2.2
faiss-cpu>=1.7.4
pandas>=2.0
openpyxl>=3.0
pdfplumber>=0.7.5
scikit-learn>=1.1
python-dotenv>=0.21
numpy==1.26.4
